cleanup(core): extract per-migration runner, output rendering, and commit helpers from migrate.ts
Move three cohesive groups of helpers out of `migrate.ts` into focused
modules. Pure refactor — no behavior change. Coverage by the migrate
test suite is preserved (294/294 passing).

- `migrate-output.ts` (new): the six presentation-layer helpers added for
  the agentic migrate UX work — per-migration boundary header, agentic
  outcome line, terminal SGR reset after agent exit, end-of-run tally
  line, inside-agent directive block, mid-run failure recap. Structural
  input types so the module stays free of cross-imports with the
  orchestrator.

- `migrate-commits.ts` (new): `commitMigrationIfRequested` and
  `commitCheckpointBeforeMigrations`. Both share a non-obvious invariant
  (compare HEAD before/after because `commitChanges` swallows `git
  commit` failures when called with a `directory`); documenting it once
  at the file head is clearer than the two copies that existed across
  `migrate.ts`.

- `agentic/run-step.ts` (new): `runAgenticPromptStep` plus its
  `AgenticPromptImplContext` / `AgenticStepResult` interfaces. Folds the
  five inner `require('./agentic/...')` calls into static imports;
  callers of `executeMigrations` lazy-require the new module so
  non-agentic runs still don't pay the cost of loading the agentic
  chain. Takes `installDepsIfChanged: () => Promise<void>` instead of
  the `ChangedDepInstaller` instance so the module is decoupled from
  the executor's internal state.

`executeMigrations` now lazy-binds the run-step entrypoint on
`agenticRun.runStep` alongside `agentic` and `runDir`, so the three call
sites (prompt-only, hybrid, generator-validation) narrow naturally
without `!` assertions.

Net: ~400 lines off `migrate.ts` (3957 → 3554) and three modules that
each have a single concern.
